Fixed the register selector and some other improvements
The essential fix was to make all general purpose registers to read data
on "high level" instead of a "rising edge". Not sure if that's a good
fix in general, but it gets me going.

Unfortunately, the RAM module stopped working reliably. Still looking
for an answer whether that's something I did wrong, or a quirk with the
Logisim ROM component itself.